0.9
Only shoot if onside and either: closest to ball, closest to ally goal and ball is in own half or ball is (roughly) in the enemy's D

If bot isn't shooting, in goal or getting boost it no longer demos enemies. Instead it puts itself in-between the ball and it's own goal, waiting for a condition to become true to take a shot or do something else

Improve reliability of fast kickoffs

If bot is going to back to goal, is offside and really close to the ball, it goes for the nearest big boost instead, to (hopefully) avoid scoring own goals.

If boost is within 1000 distance and at a greater angle than 1 bot drifts. This stops it endlessly circling boost pads

Added check which hopefully cancels shooting commands if conditions for them aren't met.

Added a link to Drive folder and in drive folder text file with links to Trello boards

Added a decal for orange team