Fast off_centre kickoff, bug fix + very_close
Added fast off_centre kickoff

Bug fix: bot would sometimes be a goalie in 1 v 1s when it's not meant to, resulting in the stack being cleared and bot doing nothing.  This was due to one of the places where checking if should be goal use closest_to_friendly_goal rather than goalie. which is only true of len(agent.friends) > 0

Added very_close variable because I wanted distances where bot moves away from ball to prevent scoring own goals and distnace where bot wouldn't cancel shot because already commited to be different. So now very_close is used for the own goal preventing and has a range of 500, while close is used for not shot cancelling with a range of 1000