Only go for kickoff if closest or joint closest and on left. Demobot fix + goalie fix. Commenting.
Instead of going for kickoff if closest or joint closest not FormularBot only goes for kickoff if closest or (joint closest and on left). It's usual in Rocket League for the player on the left to go for kickoff unles otherwise agreed.

DemoBot fix. When I moved all decision making and stack clearing code to FormularBot.py I moved code which worked out closest enemy as well. That code has now been moved back to routines.py as it's only used for demo_nearest_enemy routine and is needed by both FormularBot and DemoBot.

Fixed goalie code. In 1.3.2 I made it so goalie is false if shooting is true, however I forgot that'd make goalie rush out and shoot when ball is near enemies goal. I've now changed it to goalie is False if me_onside and (closest_to_ball or distance_ball_friendly_goal < 3000), the other two conditions of shooting.

Commented some code in FormularBot.py and DemoBot.py