Added back in demoing + other changes
Added have_possession variable which tracks if FormularBot's team last touched the ball.

Now only shoots when not(closet_to_ball) and distance_ball_foe_goal < 4000 and -1250 < agent.ball.location.x < 1250 and closest_ally_to_ball_distance > 4000
if closest ally to ball is over 3000 away from ball

if not goes for demo instead.

shooting is now not cleared if on_side and shooting, to reduce how often FormularBot decides to go for a shot and changes mind

boosts when going for demo if distance remaining is under 2000 or agent.me.boost * 100 > distance_remaining (not exact formula but somewhat acurate) rather than if boost > 50 and distance < 5000